tela pronta, responsividade feita e js feito!
