+++ view/assests/js/main.js
@@ -5,11 +5,11 @@
 // import verOngs from "./pages/verOngs";
 
 import "./components/navbar.js";
-import "./pages/criarPostagemOng.js";
 import "./pages/gestaoDePatrocinadores.js";
 import "./pages/loginUsuario.js";
 import "./pages/validacaoAtualizacao.js";
 import "./pages/validarAtualizacaoOng.js";
+import "./pages/postagemOng.js";
 
 // Arquivos que de acordo com o Luan nao existem mais por isso comentamos, 
 // porque tambemd e acordo com ele o rhyam era quem estava fazendo e nao esta presente na aula
