+++ view/pages/Ong/criarPostagemOng.php
@@ -34,4 +34,5 @@
 
     </main>
 
+    <?php require_once './../../components/footer.php' ?>
 </body>
