+++ view/pages/Ong/editarPostagemOng.php
@@ -37,33 +37,8 @@
                 </div>
             </form>
         </div>
+
     </main>
-    <script>
-        const fileInput = document.getElementById('file');
-        const preview = document.getElementById('preview');
-        const text = document.querySelector('.text');
-        const icon = document.querySelector('.icon');
-        const formPreview = document.querySelector('.formulario-imagem-preview')
 
-        fileInput.addEventListener('change', function () {
-            const file = this.files[0];
-            if (file) {
-                text.style.display = 'none';
-                icon.style.display = 'none'
-                formPreview.style.border = 'none';
-                const reader = new FileReader();
-
-                reader.addEventListener('load', function () {
-                    preview.setAttribute('src', this.result);
-                    preview.style.display = 'block';
-                });
-
-                reader.readAsDataURL(file);
-            } else {
-                preview.style.display = 'none';
-                preview.setAttribute('src', '#');
-            }
-        });
-    </script>
-
+    <?php require_once './../../components/footer.php' ?>
 </body>
